+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # Changelog
 
+## 0.6.0
+
+### Breaking changes
+
+- Changed the name of the following orchestrator options:
+    - `hideSensibleInformation` → `hideSensitiveInformation`
+    - `persistSupport` → `persistenceSupport`
+    - `persistPrefix` → `persistencePrefix`
+    - `persistEncryptionKey` → `persistenceEncryptionKey`
+
+### Fixed
+
+- `fixedInput` now works as expected.
+
 ## 0.5.4
+
+### Fixed
 
 - Fix `greedyIterate` function.
 
 ## 0.5.3
+
+### Fixed
 
 - Fix tests.
 - Update dependencies, fix `vitest` vulnerability.